+++ src/json_signatures/proofs.rs
@@ -31,11 +31,15 @@ pub const PROOF_TYPE_JCS_EIP191_LEGACY: &str ="JcsEip191Signature2022";
 // - Signature algorithm: EdDSA
 pub const PROOF_TYPE_JCS_BLAKE2_ED25519: &str = "MitraJcsEd25519Signature2022";
 
+// https://w3c.github.io/vc-di-eddsa/#jcs-eddsa-2022
+pub const CRYPTOSUITE_JCS_EDDSA: &str = "jcs-eddsa-2022";
+
 #[derive(Debug, PartialEq)]
 pub enum ProofType {
     JcsEip191Signature,
     JcsBlake2Ed25519Signature,
     JcsRsaSignature,
+    JcsEddsaSignature,
 }
 
 impl FromStr for ProofType {
@@ -58,6 +62,7 @@ impl ProofType {
     pub fn from_cryptosuite(value: &str) -> Result<Self, ConversionError> {
         let proof_type = match value {
             CRYPTOSUITE_JCS_RSA => Self::JcsRsaSignature,
+            CRYPTOSUITE_JCS_EDDSA => Self::JcsEddsaSignature,
             _ => return Err(ConversionError),
         };
         Ok(proof_type)